refactoring transaction building code
