+++ wallet/src/receiver.rs
@@ -26,7 +26,7 @@ use std::sync::{Arc, RwLock};
 use api;
 use core::consensus::reward;
 use core::core::{amount_to_hr_string, Committed, Output, Transaction, TxKernel};
-use libwallet::{aggsig, build, reward};
+use libwallet::{aggsig, build, reward, transaction};
 use core::{global, ser};
 use failure::{Fail, ResultExt};
 use keychain::{BlindingFactor, Identifier, Keychain};
@@ -63,104 +63,7 @@ fn handle_sender_initiation(
 	keychain: &Keychain,
 	partial_tx: &PartialTx,
 ) -> Result<PartialTx, Error> {
-	let (amount, _sender_pub_blinding, sender_pub_nonce, kernel_offset, _sig, tx) =
-		read_partial_tx(keychain, partial_tx)?;
-
-	let root_key_id = keychain.root_key_id();
-
-	// double check the fee amount included in the partial tx
-	// we don't necessarily want to just trust the sender
-	// we could just overwrite the fee here (but we won't) due to the ecdsa sig
-	let fee = tx_fee(
-		tx.inputs.len(),
-		tx.outputs.len() + 1,
-		tx.input_proofs_count(),
-		None,
-	);
-	if fee > tx.fee() {
-		return Err(ErrorKind::FeeDispute {
-			sender_fee: tx.fee(),
-			recipient_fee: fee,
-		})?;
-	}
-
-	if fee > amount {
-		info!(
-			LOGGER,
-			"Rejected the transfer because transaction fee ({}) exceeds received amount ({}).",
-			amount_to_hr_string(fee),
-			amount_to_hr_string(amount)
-		);
-		return Err(ErrorKind::FeeExceedsAmount {
-			sender_amount: amount,
-			recipient_fee: fee,
-		})?;
-	}
-
-	let out_amount = amount - tx.fee();
-
-	// First step is just to get the excess sum of the outputs we're participating
-	// in Output and key needs to be stored until transaction finalisation time,
-	// somehow
-
-	let key_id = WalletData::with_wallet(&config.data_file_dir, |wallet_data| {
-		let (key_id, derivation) = next_available_key(&wallet_data, keychain);
-
-		wallet_data.add_output(OutputData {
-			root_key_id: root_key_id.clone(),
-			key_id: key_id.clone(),
-			n_child: derivation,
-			value: out_amount,
-			status: OutputStatus::Unconfirmed,
-			height: 0,
-			lock_height: 0,
-			is_coinbase: false,
-			block: None,
-			merkle_proof: None,
-		});
-
-		key_id
-	})?;
-
-	// Still handy for getting the blinding sum
-	let (_, blind_sum) =
-		build::partial_transaction(vec![build::output(out_amount, key_id.clone())], keychain)
-			.context(ErrorKind::Keychain)?;
-
-	warn!(LOGGER, "Creating new aggsig context");
-	// Create a new aggsig context
-	// this will create a new blinding sum and nonce, and store them
-	let blind = blind_sum
-		.secret_key(&keychain.secp())
-		.context(ErrorKind::Keychain)?;
-	let mut context = context_manager.create_context(keychain.secp(), &partial_tx.id, blind);
-
-	context.add_output(&key_id);
-
-	let sig_part = context
-		.calculate_partial_sig(
-			keychain.secp(),
-			&sender_pub_nonce,
-			tx.fee(),
-			tx.lock_height(),
-		)
-		.unwrap();
-
-	// Build the response, which should contain sR, blinding excess xR * G, public
-	// nonce kR * G
-	let mut partial_tx = build_partial_tx(
-		&context,
-		keychain,
-		amount,
-		kernel_offset,
-		Some(sig_part),
-		tx,
-	);
-	partial_tx.phase = PartialTxPhase::ReceiverInitiation;
-
-	context_manager.save_context(context);
-
-	Ok(partial_tx)
+	transaction::recipient_initiation(config, keychain, context_manager, partial_tx)
 }
 
 /// Receive Part 3 of interactive transactions from sender, Sender Confirmation
@@ -182,73 +85,9 @@ fn handle_sender_confirmation(
 	keychain: &Keychain,
 	partial_tx: &PartialTx,
 	fluff: bool,
-) -> Result<PartialTx, Error> {
-	let (amount, sender_pub_blinding, sender_pub_nonce, kernel_offset, sender_sig_part, tx) =
-		read_partial_tx(keychain, partial_tx)?;
-	let mut context = context_manager.get_context(&partial_tx.id);
-	let sender_sig_part = sender_sig_part.unwrap();
-	let res = context.verify_partial_sig(
-		&keychain.secp(),
-		&sender_sig_part,
-		&sender_pub_nonce,
-		&sender_pub_blinding,
-		tx.fee(),
-		tx.lock_height(),
-	);
-
-	if !res {
-		error!(LOGGER, "Partial Sig from sender invalid.");
-		return Err(ErrorKind::Signature("Partial Sig from sender invalid."))?;
-	}
-
-	// Just calculate our sig part again instead of storing
-	let our_sig_part = context
-		.calculate_partial_sig(
-			&keychain.secp(),
-			&sender_pub_nonce,
-			tx.fee(),
-			tx.lock_height(),
-		)
-		.unwrap();
-
-	// And the final signature
-	let final_sig = context
-		.calculate_final_sig(
-			&keychain.secp(),
-			&sender_sig_part,
-			&our_sig_part,
-			&sender_pub_nonce,
-		)
-		.unwrap();
-
-	// Calculate the final public key (for our own sanity check)
-	let final_pubkey = context
-		.calculate_final_pubkey(&keychain.secp(), &sender_pub_blinding)
-		.unwrap();
-
-	// Check our final sig verifies
-	let res = context.verify_final_sig_build_msg(
-		&keychain.secp(),
-		&final_sig,
-		&final_pubkey,
-		tx.fee(),
-		tx.lock_height(),
-	);
-
-	if !res {
-		error!(LOGGER, "Final aggregated signature invalid.");
-		return Err(ErrorKind::Signature("Final aggregated signature invalid."))?;
-	}
-
-	let final_tx = build_final_transaction(
-		&mut context,
-		config,
-		keychain,
-		amount,
-		kernel_offset,
-		&final_sig,
-		tx.clone(),
-	)?;
+) -> Result<Transaction, Error> {
+	let final_tx =
+		transaction::recipient_confirmation(config, keychain, context_manager, partial_tx)?;
 
 	let tx_hex = to_hex(ser::ser_vec(&final_tx).unwrap());
 
@@ -263,20 +102,7 @@ fn handle_sender_confirmation(
 	}
 	api::client::post(url.as_str(), &TxWrapper { tx_hex: tx_hex }).context(ErrorKind::Node)?;
 
-	// Return what we've actually posted
-	// TODO - why build_partial_tx here? Just a naming issue?
-	let mut partial_tx = build_partial_tx(
-		&context,
-		keychain,
-		amount,
-		kernel_offset,
-		Some(final_sig),
-		tx,
-	);
-
-	context_manager.save_context(context);
-	partial_tx.phase = PartialTxPhase::ReceiverConfirmation;
-	Ok(partial_tx)
+	Ok(final_tx)
 }
 
 /// Component used to receive coins, implements all the receiving end of the
@@ -414,130 +240,4 @@ pub fn receive_coinbase(
 		reward::output(&keychain, &key_id, block_fees.fees, block_fees.height).unwrap();
 	/* .context(ErrorKind::Keychain)?; */
 	Ok((out, kern, block_fees))
-}
-
-/// builds a final transaction after the aggregated sig exchange
-fn build_final_transaction(
-	context: &mut aggsig::Context,
-	config: &WalletConfig,
-	keychain: &Keychain,
-	amount: u64,
-	kernel_offset: BlindingFactor,
-	excess_sig: &secp::Signature,
-	tx: Transaction,
-) -> Result<Transaction, Error> {
-	let root_key_id = keychain.root_key_id();
-
-	// double check the fee amount included in the partial tx
-	// we don't necessarily want to just trust the sender
-	// we could just overwrite the fee here (but we won't) due to the ecdsa sig
-	let fee = tx_fee(
-		tx.inputs.len(),
-		tx.outputs.len() + 1,
-		tx.input_proofs_count(),
-		None,
-	);
-	if fee > tx.fee() {
-		return Err(ErrorKind::FeeDispute {
-			sender_fee: tx.fee(),
-			recipient_fee: fee,
-		})?;
-	}
-
-	if fee > amount {
-		info!(
-			LOGGER,
-			"Rejected the transfer because transaction fee ({}) exceeds received amount ({}).",
-			amount_to_hr_string(fee),
-			amount_to_hr_string(amount)
-		);
-		return Err(ErrorKind::FeeExceedsAmount {
-			sender_amount: amount,
-			recipient_fee: fee,
-		})?;
-	}
-
-	let out_amount = amount - tx.fee();
-
-	// Get output we created in earlier step
-	// TODO: will just be one for now, support multiple later
-	let output_vec = context.get_outputs();
-
-	// operate within a lock on wallet data
-	let (key_id, derivation) = WalletData::with_wallet(&config.data_file_dir, |wallet_data| {
-		let (key_id, derivation) = retrieve_existing_key(&wallet_data, output_vec[0].clone());
-
-		wallet_data.add_output(OutputData {
-			root_key_id: root_key_id.clone(),
-			key_id: key_id.clone(),
-			n_child: derivation,
-			value: out_amount,
-			status: OutputStatus::Unconfirmed,
-			height: 0,
-			lock_height: 0,
-			is_coinbase: false,
-			block: None,
-			merkle_proof: None,
-		});
-
-		(key_id, derivation)
-	})?;
-
-	// Build final transaction, the sum of which should
-	// be the same as the exchanged excess values
-	let mut final_tx = build::transaction(
-		vec![
-			build::initial_tx(tx),
-			build::output(out_amount, key_id.clone()),
-			build::with_offset(kernel_offset),
-		],
-		keychain,
-	).context(ErrorKind::Keychain)?;
-
-	// build the final excess based on final tx and offset
-	let final_excess = {
-		// TODO - do we need to verify rangeproofs here?
-		for x in &final_tx.outputs {
-			x.verify_proof().context(ErrorKind::Transaction)?;
-		}
-
-		// sum the input/output commitments on the final tx
-		let overage = final_tx.fee() as i64;
-		let tx_excess = final_tx
-			.sum_commitments(overage, None)
-			.context(ErrorKind::Transaction)?;
-
-		// subtract the kernel_excess (built from kernel_offset)
-		let offset_excess = keychain
-			.secp()
-			.commit(0, kernel_offset.secret_key(&keychain.secp()).unwrap())
-			.unwrap();
-		keychain
-			.secp()
-			.commit_sum(vec![tx_excess], vec![offset_excess])
-			.context(ErrorKind::Transaction)?
-	};
-
-	// update the tx kernel to reflect the offset excess and sig
-	assert_eq!(final_tx.kernels.len(), 1);
-	final_tx.kernels[0].excess = final_excess.clone();
-	final_tx.kernels[0].excess_sig = excess_sig.clone();
-
-	// confirm the kernel verifies successfully before proceeding
-	final_tx.kernels[0]
-		.verify()
-		.context(ErrorKind::Transaction)?;
-
-	// confirm the overall transaction is valid (including the updated kernel)
-	let _ = final_tx.validate().context(ErrorKind::Transaction)?;
-
-	debug!(
-		LOGGER,
-		"Finalized transaction and built output - {:?}, {:?}, {}",
-		root_key_id.clone(),
-		key_id.clone(),
-		derivation,
-	);
-
-	Ok(final_tx)
 }
