+++ wallet/src/types.rs
@@ -767,6 +767,7 @@ pub struct PartialTx {
 	pub phase: PartialTxPhase,
 	pub id: Uuid,
 	pub amount: u64,
+	pub lock_height: u64,
 	pub public_blind_excess: String,
 	pub public_nonce: String,
 	pub kernel_offset: String,
@@ -781,6 +782,7 @@ pub fn build_partial_tx(
 	context: &aggsig::Context,
 	keychain: &keychain::Keychain,
 	receive_amount: u64,
+	lock_height: u64,
 	kernel_offset: BlindingFactor,
 	part_sig: Option<secp::Signature>,
 	tx: Transaction,
@@ -798,6 +800,7 @@ pub fn build_partial_tx(
 		phase: PartialTxPhase::SenderInitiation,
 		id: context.transaction_id,
 		amount: receive_amount,
+		lock_height: lock_height,
 		public_blind_excess: util::to_hex(pub_excess),
 		public_nonce: util::to_hex(pub_nonce),
 		kernel_offset: kernel_offset.to_hex(),
@@ -816,6 +819,7 @@ pub fn read_partial_tx(
 	partial_tx: &PartialTx,
 ) -> Result<
 	(
+		u64,
 		u64,
 		PublicKey,
 		PublicKey,
@@ -850,7 +854,15 @@ pub fn read_partial_tx(
 	let tx = ser::deserialize(&mut &tx_bin[..]).context(ErrorKind::GenericError(
 		"Could not deserialize transaction, invalid format.",
 	))?;
-	Ok((partial_tx.amount, blinding, nonce, kernel_offset, sig, tx))
+	Ok((
+		partial_tx.amount,
+		partial_tx.lock_height,
+		blinding,
+		nonce,
+		kernel_offset,
+		sig,
+		tx,
+	))
 }
 
 /// Amount in request to build a coinbase output.
